Refactor evaluator into dedicated components
